+++ platform/switch/source/objects/source.cpp
@@ -8,15 +8,15 @@ using namespace love;
 
 #define AudioModule() (Module::GetInstance<Audio>(Module::M_AUDIO))
 
-StaticDataBuffer::StaticDataBuffer(void * data, size_t size)
+StaticDataBuffer::StaticDataBuffer(void* data, size_t size)
 {
-    std::pair<void *, size_t> buff = AudioPool::MemoryAlign(size);
+    std::pair<void*, size_t> buff = AudioPool::MemoryAlign(size);
 
     if (buff.first)
     {
         memcpy(buff.first, data, size);
 
-        this->buffer = {(s16 *)buff.first, buff.second};
+        this->buffer = { (s16*)buff.first, buff.second };
     }
 }
 
@@ -27,20 +27,20 @@ StaticDataBuffer::~StaticDataBuffer()
 
 /* SOURCE IMPLEMENTATION */
 
-Source::Source(Pool * pool, SoundData * sound) : common::Source(pool, sound)
+Source::Source(Pool* pool, SoundData* sound) : common::Source(pool, sound)
 {
-    this->sources[0] = AudioDriverWaveBuf();
+    this->sources[0]                     = AudioDriverWaveBuf();
     this->sources[0].start_sample_offset = 0;
-    this->sources[0].size = sound->GetSize();
-    this->sources[0].end_sample_offset = sound->GetSampleCount();
+    this->sources[0].size                = sound->GetSize();
+    this->sources[0].end_sample_offset   = sound->GetSampleCount();
 }
 
-Source::Source(Pool * pool, Decoder * decoder) : common::Source(pool, decoder)
+Source::Source(Pool* pool, Decoder* decoder) : common::Source(pool, decoder)
 {
     this->InitializeStreamBuffers(decoder);
 }
 
-love::Source * Source::Clone()
+love::Source* Source::Clone()
 {
     return new Source(*this);
 }
@@ -51,15 +51,15 @@ Source::~Source()
     this->FreeBuffer();
 }
 
-void Source::InitializeStreamBuffers(Decoder * decoder)
+void Source::InitializeStreamBuffers(Decoder* decoder)
 {
-    for (auto & source : this->sources)
+    for (auto& source : this->sources)
     {
-        source = AudioDriverWaveBuf();
+        source                     = AudioDriverWaveBuf();
         source.start_sample_offset = 0;
-        source.size = 0;
-        source.data_pcm16 = (s16 *)AudioPool::MemoryAlign(decoder->GetSize()).first;
-        source.state = AudioDriverWaveBufState_Done;
+        source.size                = 0;
+        source.data_pcm16          = (s16*)AudioPool::MemoryAlign(decoder->GetSize()).first;
+        source.state               = AudioDriverWaveBufState_Done;
     }
 }
 
@@ -68,15 +68,16 @@ void Source::FreeBuffer()
     if (this->sourceType != TYPE_STATIC)
     {
         for (int i = 0; i < Source::MAX_BUFFERS; i++)
-            AudioPool::MemoryFree(std::make_pair(this->sources[i].data_pcm16, this->sources[i].size));
+            AudioPool::MemoryFree(
+                std::make_pair(this->sources[i].data_pcm16, this->sources[i].size));
     }
 }
 
-int Source::GetSampleOffset()
+double Source::GetSampleOffset()
 {
-    int offset = 0;
+    double offset = 0;
 
-    AudioModule()->GetDriver()->LockFunction([this, &offset](AudioDriver * driver) {
+    AudioModule()->GetDriver()->LockFunction([this, &offset](AudioDriver* driver) {
         offset = audrvVoiceGetPlayedSampleCount(driver, this->channel);
     });
 
@@ -85,7 +86,7 @@ int Source::GetSampleOffset()
 
 void Source::SetVolume(float volume)
 {
-    AudioModule()->GetDriver()->LockFunction([this, volume](AudioDriver * driver) {
+    AudioModule()->GetDriver()->LockFunction([this, volume](AudioDriver* driver) {
         audrvVoiceSetVolume(driver, this->channel, volume);
     });
 
@@ -94,7 +95,7 @@ void Source::SetVolume(float volume)
 
 void Source::Reset()
 {
-    AudioModule()->GetDriver()->LockFunction([this](AudioDriver * driver) {
+    AudioModule()->GetDriver()->LockFunction([this](AudioDriver* driver) {
         audrvVoiceInit(driver, this->channel, this->channels, PcmFormat_Int16, this->sampleRate);
         audrvVoiceSetDestinationMix(driver, this->channel, AUDREN_FINAL_MIX_ID);
 
@@ -126,7 +127,7 @@ bool Source::Update()
                     if (decoded == 0)
                         return false;
 
-                    AudioModule()->GetDriver()->LockFunction([this, which](AudioDriver * driver) {
+                    AudioModule()->GetDriver()->LockFunction([this, which](AudioDriver* driver) {
                         audrvVoiceAddWaveBuf(driver, this->channel, &this->sources[which]);
                         audrvVoiceStart(driver, this->channel);
                     });
@@ -152,7 +153,7 @@ void Source::PrepareAtomic()
     switch (this->sourceType)
     {
         case TYPE_STATIC:
-            this->sources[0].data_pcm16 = (s16 *)this->staticBuffer.Get()->GetBuffer();
+            this->sources[0].data_pcm16 = (s16*)this->staticBuffer.Get()->GetBuffer();
             break;
         case TYPE_STREAM:
         {
@@ -182,7 +183,8 @@ int Source::StreamAtomic(size_t which)
         this->decoder->Rewind();
 
     this->sources[which].size = decoded;
-    this->sources[which].end_sample_offset = (int)((decoded / this->channels) / (this->bitDepth / 8));
+    this->sources[which].end_sample_offset =
+        (int)((decoded / this->channels) / (this->bitDepth / 8));
 
     return decoded;
 }
@@ -198,7 +200,8 @@ bool Source::IsPlaying() const
     size_t bufferCount = (this->sourceType == TYPE_STREAM) ? MAX_BUFFERS : 1;
     /* check if any of our sources are queued or playing */
     for (size_t index = 0; index < bufferCount; index++)
-        sourcePlaying |= (this->sources[index].state == AudioDriverWaveBufState_Queued || this->sources[index].state == AudioDriverWaveBufState_Playing);
+        sourcePlaying |= (this->sources[index].state == AudioDriverWaveBufState_Queued ||
+                          this->sources[index].state == AudioDriverWaveBufState_Playing);
 
     return sourcePlaying;
 }
@@ -213,7 +216,7 @@ bool Source::IsFinished() const
 
     bool finished = false;
 
-    AudioModule()->GetDriver()->LockFunction([this, &finished](AudioDriver * driver) {
+    AudioModule()->GetDriver()->LockFunction([this, &finished](AudioDriver* driver) {
         finished = (audrvVoiceIsPlaying(driver, this->channel) == false);
     });
 
@@ -232,14 +235,14 @@ bool Source::PlayAtomic()
         armDCacheFlush(this->sources[0].data_pcm16, this->staticBuffer->GetSize());
 
     /* add the initial wavebuffer */
-    AudioModule()->GetDriver()->LockFunction([this](AudioDriver * driver) {
+    AudioModule()->GetDriver()->LockFunction([this](AudioDriver* driver) {
         audrvVoiceAddWaveBuf(driver, this->channel, &this->sources[0]);
         audrvVoiceStart(driver, this->channel);
     });
 
     success = true;
 
-    //isPlaying() needs source to be valid
+    // isPlaying() needs source to be valid
     if (this->sourceType == TYPE_STREAM)
     {
         this->valid = true;
@@ -248,7 +251,7 @@ bool Source::PlayAtomic()
             success = false;
     }
 
-    //stop() needs source to be valid
+    // stop() needs source to be valid
     if (!success)
     {
         this->valid = true;
@@ -263,7 +266,7 @@ bool Source::PlayAtomic()
 
 void Source::PauseAtomic()
 {
-    AudioModule()->GetDriver()->LockFunction([this](AudioDriver * driver) {
+    AudioModule()->GetDriver()->LockFunction([this](AudioDriver* driver) {
         if (this->valid)
             audrvVoiceSetPaused(driver, this->channel, true);
     });
@@ -273,15 +276,14 @@ void Source::ResumeAtomic()
 {
     if (this->valid && !this->IsPlaying())
     {
-        AudioModule()->GetDriver()->LockFunction([this](AudioDriver * driver) {
-            audrvVoiceSetPaused(driver, this->channel, false);
-        });
+        AudioModule()->GetDriver()->LockFunction(
+            [this](AudioDriver* driver) { audrvVoiceSetPaused(driver, this->channel, false); });
     }
 }
 
 void Source::StopAtomic()
 {
-    AudioModule()->GetDriver()->LockFunction([this](AudioDriver * driver) {
+    AudioModule()->GetDriver()->LockFunction([this](AudioDriver* driver) {
         if (!this->valid)
             return;
